v27.0.2: command-driven TOML architecture, 6 new widgets, Ctx improvements
- Add src/framework/command.rs with CommandRunner, BoundCommand, OutputParser, AppConfig, WidgetConfig
- Add Ctx::run_command() and Ctx::available_commands() for AI command enumeration
- Add 6 new widgets: ConfirmDialog, KeyValueGrid, Gauge, LogViewer, StreamingText, StatusBadge
- Add toml=0.8 dependency for TOML-driven widget configuration
- Extend Widget trait with commands() method for AI surface
- Add App::from_toml(), App::add_command(), App::run_command(), App::available_commands()
- All 650+ tests passing, clippy warnings cleaned

+++ src/framework/layout.rs
@@ -134,7 +134,7 @@ impl Layout {
         let ratio_total: u32 = ratios.iter().map(|(_, n, _)| *n as u32).sum();
         for (i, n, d) in ratios.iter() {
             if *d > 0 && ratio_total > 0 {
-                let size = (remaining as u32 * *n as u32 / ratio_total as u32) as u16;
+                let size = (remaining as u32 * *n as u32 / ratio_total) as u16;
                 sizes[*i] = sizes[*i].max(size);
             }
         }
